requirements.txt
streamlit
tensorflow
numpy
matplotlib
seaborn
scikit-learn
Pillow
